+++ dynamic_color/lib/src/dynamic_color_builder.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:material_color_utilities/material_color_utilities.dart';
 
+import 'corepalette_to_colorscheme.dart';
 import 'dynamic_color_plugin.dart';
 
 /// A stateful builder widget that provides a light and dark [ColorScheme].
@@ -70,53 +71,5 @@ class _DynamicColorBuilderState extends State<DynamicColorBuilder> {
   @override
   Widget build(BuildContext context) {
     return widget.builder(_light, _dark);
-  }
-}
-
-extension on CorePalette {
-  /// Create a [ColorScheme] from the given `palette` obtained from the Android OS.
-  ColorScheme? toColorScheme({
-    Brightness brightness = Brightness.light,
-  }) {
-    final Scheme scheme;
-
-    switch (brightness) {
-      case Brightness.light:
-        scheme = Scheme.lightFromCorePalette(this);
-        break;
-      case Brightness.dark:
-        scheme = Scheme.darkFromCorePalette(this);
-        break;
-    }
-    return ColorScheme(
-      primary: Color(scheme.primary),
-      onPrimary: Color(scheme.onPrimary),
-      primaryContainer: Color(scheme.primaryContainer),
-      onPrimaryContainer: Color(scheme.onPrimaryContainer),
-      secondary: Color(scheme.secondary),
-      onSecondary: Color(scheme.onSecondary),
-      secondaryContainer: Color(scheme.secondaryContainer),
-      onSecondaryContainer: Color(scheme.onSecondaryContainer),
-      tertiary: Color(scheme.tertiary),
-      onTertiary: Color(scheme.onTertiary),
-      tertiaryContainer: Color(scheme.tertiaryContainer),
-      onTertiaryContainer: Color(scheme.onTertiaryContainer),
-      error: Color(scheme.error),
-      onError: Color(scheme.onError),
-      errorContainer: Color(scheme.errorContainer),
-      onErrorContainer: Color(scheme.onErrorContainer),
-      outline: Color(scheme.outline),
-      background: Color(scheme.background),
-      onBackground: Color(scheme.onBackground),
-      surface: Color(scheme.surface),
-      onSurface: Color(scheme.onSurface),
-      surfaceVariant: Color(scheme.surfaceVariant),
-      onSurfaceVariant: Color(scheme.onSurfaceVariant),
-      inverseSurface: Color(scheme.inverseSurface),
-      onInverseSurface: Color(scheme.inverseOnSurface),
-      inversePrimary: Color(scheme.inversePrimary),
-      shadow: Color(scheme.shadow),
-      brightness: brightness,
-    );
   }
 }